refactor: centralize symmetric LM-head provider lifecycle
