+++ lmdeploy/pytorch/nn/embedding.py
@@ -1,12 +1,10 @@
 # Copyright (c) OpenMMLab. All rights reserved.
-import logging
-
 import torch
 import torch.distributed as dist
 from torch import nn
 
-from lmdeploy.pytorch import envs as _envs
 from lmdeploy.pytorch.backends import get_backend
+from lmdeploy.pytorch.backends.cuda.comm.symm_mem_allgather import MultimemAllGatherer
 from lmdeploy.pytorch.backends.embedding import EmbeddingBuildSpec
 from lmdeploy.pytorch.backends.linear import LinearBuildSpec
 from lmdeploy.pytorch.distributed import get_dist_group, get_dist_manager, get_tp_world_rank
@@ -14,23 +12,6 @@ from lmdeploy.pytorch.models.patch import get_build_model_context
 from lmdeploy.pytorch.weight_loader.model_weight_loader import default_weight_loader
 
 DEFAULT_VOCAB_PADDING_SIZE = 64
-logger = logging.getLogger(__name__)
-
-
-def _tp_agree(local_ready: bool, device: torch.device, group: dist.ProcessGroup) -> bool:
-    """Resolve an optional LM-head provider decision on every TP rank."""
-    ready = torch.tensor(int(local_ready), dtype=torch.int32, device=device)
-    dist.all_reduce(ready, op=dist.ReduceOp.MIN, group=group)
-    return bool(ready.item())
-
-
-def _tp_same_config(values: tuple[int, ...], device: torch.device, group: dist.ProcessGroup) -> bool:
-    """Return whether every TP rank supplied the same integer config."""
-    lower = torch.tensor(values, dtype=torch.int64, device=device)
-    upper = lower.clone()
-    dist.all_reduce(lower, op=dist.ReduceOp.MIN, group=group)
-    dist.all_reduce(upper, op=dist.ReduceOp.MAX, group=group)
-    return bool(torch.equal(lower, upper))
 
 
 def pad_vocab_size(vocab_size: int, pad_to: int = DEFAULT_VOCAB_PADDING_SIZE) -> int:
@@ -173,84 +154,21 @@ class ParallelLMHead(ParallelEmbedding):
             enable_deterministic=get_build_model_context().enable_deterministic,
         )
 
-        self._symm_mem_gatherer = None
-        self._symm_mem_device = self.weight.device
-        self._symm_mem_dtype = self.weight.dtype
-        self._logged_nccl_small_token_fallback = False
-        if self.all_reduce and self.weight.device.type == 'cuda':
-            device = self.weight.device
-            if device.index is None:
-                device = torch.device('cuda', torch.cuda.current_device())
-            requested = _envs.enable_symm_mem_lmhead and self.weight.dtype == torch.bfloat16
-            if not _tp_agree(requested, device, self.tp_group):
-                return
-
-            gathered_width = self.tp * self.vocab_size_padded
-            capacity = _envs.symm_mem_lmhead_max_mb * 1024 * 1024
-            max_tokens = capacity // (gathered_width * torch.bfloat16.itemsize)
-            barrier_mode = {'auto': 0, 'per_block': 1, 'single': 2}.get(
-                _envs.symm_mem_lmhead_barrier_mode, -1)
-            same_config = _tp_same_config(
-                (capacity, gathered_width, max_tokens,
-                 _envs.symm_mem_lmhead_blocks,
-                 _envs.symm_mem_lmhead_block_threads,
-                 int(_envs.symm_mem_lmhead_autotune), barrier_mode,
-                 _envs.symm_mem_lmhead_min_tokens), device, self.tp_group)
-            if max_tokens <= 0 or not same_config:
-                if self.tp_rank == 0:
-                    logger.warning('symmetric-memory LM-head disabled because TP ranks have inconsistent arena config')
-                return
-
-            gatherer_cls = None
-            try:
-                from lmdeploy.pytorch.backends.cuda.comm.symm_mem_allgather import MultimemAllGatherer
-                gatherer_cls = MultimemAllGatherer
-            except ImportError as exc:
-                if self.tp_rank == 0:
-                    logger.warning('symmetric-memory LM-head unavailable: %s', exc)
-            if not _tp_agree(gatherer_cls is not None, device, self.tp_group):
-                return
-
-            gatherer = gatherer_cls(group=self.tp_group,
-                                    rank=self.tp_rank,
-                                    gathered_width=gathered_width,
-                                    max_tokens=max_tokens)
-            prepared = gatherer.prepare(device)
-            admit_static = getattr(gatherer, 'admit_static', None)
-            if prepared and callable(admit_static):
-                # Freeze the static BF16/layout contract during model setup;
-                # the first decode call can then dispatch directly without
-                # another TP-wide admission reduction.
-                prepared = admit_static(device)
-            if prepared:
-                self._symm_mem_gatherer = gatherer
+        self._symm_mem_gatherer = (
+            MultimemAllGatherer(self.tp_group, self.tp_rank,
+                               self.tp * self.vocab_size_padded,
+                               self.weight.device, self.weight.dtype)
+            if self.all_reduce else None)
 
     def tie_weights(self, embedding: ParallelEmbedding):
         """Tie the local LM-head shard to a parallel embedding shard."""
         self.weight = embedding.weight
 
     def _apply(self, fn, recurse=True):
-        """Keep the symmetric arena aligned with model device moves."""
-        previous_device = self._symm_mem_device
-        previous_dtype = self._symm_mem_dtype
+        """Notify the provider after coordinated model device/dtype moves."""
         result = super()._apply(fn, recurse=recurse)
-        current_device = self.weight.device
-        current_dtype = self.weight.dtype
-        gatherer = self._symm_mem_gatherer
-        if gatherer is not None and (current_device != previous_device or current_dtype != previous_dtype):
-            if previous_device.type == 'cuda':
-                gatherer.release()
-            self._symm_mem_device = current_device
-            self._symm_mem_dtype = current_dtype
-            if current_dtype != torch.bfloat16:
-                self._symm_mem_gatherer = None
-            elif current_device.type == 'cuda':
-                prepared = gatherer.prepare(current_device)
-                admit_static = getattr(gatherer, 'admit_static', None)
-                if prepared and callable(admit_static):
-                    prepared = admit_static(current_device)
-                if not prepared:
-                    self._symm_mem_gatherer = None
+        if self._symm_mem_gatherer is not None:
+            self._symm_mem_gatherer.reset_for_weight(self.weight)
         return result
 
     def get_local_logits(self, hidden_states: torch.Tensor):
@@ -265,25 +183,10 @@ class ParallelLMHead(ParallelEmbedding):
             return local_logits[..., :self.vocab_size]
 
         if self._symm_mem_gatherer is not None:
-            local_logits_2d = local_logits.reshape(-1, local_logits.shape[-1])
-            if local_logits_2d.shape[0] >= _envs.symm_mem_lmhead_min_tokens:
-                # ParallelLMHead has a stable BF16/layout contract after the
-                # first collective admission.  Use the admitted fast path so
-                # steady-state decode avoids repeated host-side validation;
-                # arbitrary gatherer users retain the defensive __call__ API.
-                fast_call = getattr(self._symm_mem_gatherer, 'fast_call', None)
-                gathered = (fast_call(local_logits_2d)
-                            if callable(fast_call) else
-                            self._symm_mem_gatherer(local_logits_2d))
-                if gathered is not None:
-                    output_shape = local_logits.shape[:-1] + (self.tp * local_logits.shape[-1], )
-                    return gathered.reshape(output_shape)[..., :self.vocab_size]
-            elif not self._logged_nccl_small_token_fallback:
-                logger.info(
-                    'symmetric-memory LM-head fallback to NCCL for token_rows=%s '
-                    '(minimum=%s)', local_logits_2d.shape[0],
-                    _envs.symm_mem_lmhead_min_tokens)
-                self._logged_nccl_small_token_fallback = True
+            gathered = self._symm_mem_gatherer(local_logits.reshape(-1, local_logits.shape[-1]))
+            if gathered is not None:
+                output_shape = local_logits.shape[:-1] + (self.tp * local_logits.shape[-1], )
+                return gathered.reshape(output_shape)[..., :self.vocab_size]
 
         input_size = local_logits.size()
         output_size = (input_size[0] * self.tp, ) + input_size[1:]
